[FIX] correct shapes in wrapper, and dataset (needs to be checked) pass logging of images, add mask and plucer to VanillaCFG

+++ sgm/modules/diffusionmodules/denoiser.py
@@ -35,7 +35,7 @@ class Denoiser(nn.Module):
         c_noise = self.possibly_quantize_c_noise(c_noise.reshape(sigma_shape))
         
         if "mask" in cond:
-            mask = cond.pop("mask")[...,None,None,None]
+            mask = cond.pop("mask")[...,None,None,None].to(dtype=input.dtype)
             input = input * (1 - mask) + input * mask
             
         return (
